requirements.txt
aiocontextvars==0.2.2
aiohttp==3.6.2
aiohttp-cors==0.7.0
aiohttp-jinja2==1.2.0
alabaster==0.7.12
aniso8601==8.0.0
apispec==3.1.0
argh==0.26.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
asyncmock==0.4.1
asyncpg==0.20.0
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.7.0
blinker==1.4
blockdiag==1.5.4
bump2version==1.0.0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
codecov==2.0.15
contextvars==2.4
coverage==4.5.4
docutils==0.15.2
filtrark==0.9.0
funcparserlib==0.3.6
future==0.18.2
idna==2.8
idna-ssl==1.1.0
imagesize==1.1.0
immutables==0.14
importlib-metadata==0.23
injectark==0.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
lazy-object-proxy==1.4.3
livereload==2.6.1
MarkupSafe==1.1.1
marshmallow==3.2.2
mccabe==0.6.1
migrark==0.4.1
mock==3.0.5
more-itertools==7.2.0
multidict==4.7.5
mypy==0.740
mypy-extensions==0.4.3
packaging==19.2
passlib==1.7.1
pathlib==1.0.1
pathtools==0.1.2
pep8==1.7.1
Pillow==6.2.1
pluggy==0.13.0
port-for==0.3.1
psycopg2==2.8.4
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
PyNaCl==1.3.0
pyparsing==2.4.5
pyrsistent==0.15.5
pytest==5.2.4
pytest-aiohttp==0.3.0
pytest-asyncio==0.10.0
pytest-cov==2.7.1
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-mock==1.10.2
pytest-randomly==3.0.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.0
python-rapidjson==0.9.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
rope==0.17.0
semver==2.9.0
seqdiag==0.9.6
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinx-autobuild==0.7.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-seqdiag==0.8.5
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
tenark==0.6.4
tornado==6.0.3
typed-ast==1.4.0
typing-extensions==3.7.4.1
ujson==1.35
urllib3==1.25.7
urwid==2.1.0
uvloop==0.14.0
watchdog==0.9.0
wcwidth==0.1.7
webcolors==1.10
Werkzeug==0.16.0
widark==0.2.6
wrapt==1.11.2
yarl==1.4.2
zipp==0.6.0
